% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{Bagging@fit_bag}
\alias{Bagging@fit_bag}
\title{Fit the Bagging ensemble to training data}
\arguments{
\item{bagging}{A Bagging object}

\item{X}{Matrix of predictors}

\item{y}{Numeric target vector}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Fit the Bagging ensemble to training data
}
\examples{
fit_bag(bag, X, y)
}
