% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{Bagging@get_tree}
\alias{Bagging@get_tree}
\title{Get a single tree from the bagged ensemble}
\arguments{
\item{bagging}{A Bagging object}

\item{b}{Integer index of the tree}
}
\value{
The \code{CART} object stored at index \code{b}
}
\description{
Get a single tree from the bagged ensemble
}
\examples{
get_tree(bagging, 1L)
}
