% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{Bagging@predict_bag}
\alias{Bagging@predict_bag}
\title{Predict using the Bagging ensemble}
\arguments{
\item{bagging}{A Bagging object}

\item{x}{Numeric vector representing one observation}
}
\value{
Numeric prediction
}
\description{
Predict using the Bagging ensemble
}
\examples{
predict_bag(bag, c(0.1, 0.2))
}
