% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@assign_right_child}
\alias{BinaryTree@assign_right_child}
\title{Assign right child}
\arguments{
\item{tree}{A BinaryTree object}

\item{parent}{A Node object}

\item{child}{A Node object}
}
\value{
void
}
\description{
Assign right child
}
\examples{
assign_right_child(tree,	parent,	child)
}
