% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@count_leaves}
\alias{BinaryTree@count_leaves}
\title{Counts the number of leaves in a tree}
\arguments{
\item{tree}{A BinaryTree object}
}
\value{
Integer number of leaves
}
\description{
Counts the number of leaves in a tree
}
\examples{
count_leaves(tree)
}
