% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@get_depth}
\alias{BinaryTree@get_depth}
\title{Getter function for tree depth}
\arguments{
\item{tree}{A BinaryTree object}
}
\value{
depth	Integer value
}
\description{
Getter function for tree depth
}
\examples{
get_depth(tree)
}
