% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@is_empty}
\alias{BinaryTree@is_empty}
\title{Verify whether tree is empty}
\arguments{
\item{tree}{BinaryTree object}
}
\value{
True/False
}
\description{
Verify whether tree is empty
}
\examples{
is_empty(tree)
}
