% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@predict_tree}
\alias{BinaryTree@predict_tree}
\title{Predict the output for a single observation}
\arguments{
\item{tree}{A BinaryTree object}

\item{x}{A numeric vector of length d (one observation)}
}
\value{
Double corresponding to the leaf value reached by x
}
\description{
Predict the output for a single observation
}
\examples{
predict_tree(tree, x)
}
