% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@prune}
\alias{BinaryTree@prune}
\title{Prune a node, collapsing it and its subtree back into a leaf}
\arguments{
\item{tree}{A BinaryTree object}

\item{node}{A Node object to prune (must be an inner node)}

\item{leaf_value}{A double to assign as the new leaf value}
}
\value{
void
}
\description{
Prune a node, collapsing it and its subtree back into a leaf
}
\examples{
prune(tree, node, leaf_value)
}
