% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@update_depth}
\alias{BinaryTree@update_depth}
\title{Increment tree depth}
\arguments{
\item{tree}{A BinaryTree object}
}
\value{
void
}
\description{
Increment tree depth
}
\examples{
update_depth(tree)
}
