% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@update_n_leaves}
\alias{BinaryTree@update_n_leaves}
\title{Increment tree number of leaves}
\arguments{
\item{tree}{A BinaryTree object}
}
\value{
void
}
\description{
Increment tree number of leaves
}
\examples{
update_n_leaves(tree)
}
