% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_tree.R
\name{BinaryTree@update_n_nodes}
\alias{BinaryTree@update_n_nodes}
\title{Increment tree number of nodes}
\arguments{
\item{tree}{A BinaryTree object}
}
\value{
void
}
\description{
Increment tree number of nodes
}
\examples{
update_n_nodes(tree)
}
