% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{CART@build_tree}
\alias{CART@build_tree}
\title{Recursively build the CART tree}
\arguments{
\item{cart}{A CART object}

\item{node}{A Node object}

\item{X}{A numeric matrix of features (n    x    d)}

\item{y}{A numeric vector of target values}

\item{depth}{Current depth integer}
}
\value{
void
}
\description{
Recursively build the CART tree
}
\examples{
build_tree(cart,	node,	X,	y,	depth)
}
