% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{CART@compute_leaf_value}
\alias{CART@compute_leaf_value}
\title{Compute	the leaf value for a node}
\arguments{
\item{cart}{A CART object}

\item{y}{A numeric vector of target values}
}
\value{
Double	(regression: mean) or integer (classification: majority class)
}
\description{
Compute	the leaf value for a node
}
\examples{
compute_leaf_value(cart,	y)
}
