% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{CART@cost_complexity_prune}
\alias{CART@cost_complexity_prune}
\title{Perform cost-complexity pruning on a CART tree}
\arguments{
\item{cart}{A CART object}

\item{X}{Numeric feature matrix}

\item{y}{Numeric response vector}
}
\value{
A list containing
\itemize{
\item trees: List of pruned CART trees
\item ratios: Numeric vector of cost-complexity parameters
}
}
}
\description{
Starts from a fully grown tree, repeatedly removes the weakest links.
}
\details{
The result is a sequence of subtrees from Satz 6.19
together with their corresponding complexity ratios.

Trees are deep-copied using serialization in order to store
intermediate pruning states, since the underlying data structures use environments.
}
\examples{
pruned <- cost_complexity_prune(cart, X, y)
pruned$trees
pruned$ratios
}
