% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{CART@find_split}
\alias{CART@find_split}
\title{Find the best split (j,	s) for a node given data subset}
\arguments{
\item{cart}{A CART object}

\item{X}{A numeric matrix of features (n    x    d)}

\item{y}{A numeric vector of target values}
}
\value{
A list with j (feature	index) and s (split	point), or NULL if no valid split
}
\description{
Find the best split (j,	s) for a node given data subset
}
\examples{
find_split(cart,	X,	y)
}
