% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{CART@find_weakest_link}
\alias{CART@find_weakest_link}
\title{Find the weakest link in a CART tree}
\arguments{
\item{cart}{A CART object}

\item{X}{Numeric feature matrix used for training}

\item{y}{Numeric response vector}
}
\value{
A list containing
\itemize{
\item A Node object: the weakest link
\item Numeric: cc_ratio the minimum cost-complexity ratio
}
}
\description{
Identifies the internal node whose removal results in the smallest cost-complexity ratio.
}
\examples{
find_weakest_link(cart, X, y)
}
