% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{CART@fit}
\alias{CART@fit}
\title{Fit the CART to training data}
\arguments{
\item{cart}{A CART object}

\item{X}{A numeric matrix of features (n    x    d)}

\item{y}{A numeric vector of target values}
}
\value{
void
}
\description{
Fit the CART to training data
}
\examples{
fit(cart,	X,	y)
}
