% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{compute_cost_complexity_ratio}
\alias{compute_cost_complexity_ratio}
\title{Compute the cost-complexity ratio for a node}
\usage{
compute_cost_complexity_ratio(node, root, X, y, type)
}
\arguments{
\item{node}{Node whose pruning cost is evaluated}

\item{root}{Root node of the tree}

\item{X}{Numeric feature matrix}

\item{y}{Numeric target vector}

\item{type}{Tree type ("regression" or "classification")}
}
\value{
Numeric cost-complexity ratio
}
\description{
Computes the factor  (R(t) - R(T_t))/(#T_t - 1) from Satz 6.18/6.19, where
\itemize{
\item R(t) error if the subtree rooted at t gets replaced by leaf
\item R(T_t) total error in the node's subtree
\item #T_t number of leaves in the node's subtree
}
}
\examples{
compute_cost_complexity_ratio(node, cart@ref$root, X, y, "regression")
}
