% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{compute_leaf_prediction}
\alias{compute_leaf_prediction}
\title{Compute a leaf prediction from target values}
\usage{
compute_leaf_prediction(values, type)
}
\arguments{
\item{values}{Numeric vector of response values}

\item{type}{"regression" or "classification"}
}
\value{
Numeric predicted leaf value.
}
\description{
regression: mean
classification: majority class
}
\examples{
compute_leaf_prediction(y, "regression")
}
