% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{count_subtree_leaves}
\alias{count_subtree_leaves}
\title{Count the number of leaves in a subtree rooted at given node}
\usage{
count_subtree_leaves(node)
}
\arguments{
\item{node}{A Node object}
}
\value{
Integer number of leaves
}
\description{
Count the number of leaves in a subtree rooted at given node
}
\examples{
count_subtree_leaves(node)
}
