% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{get_internal_nodes}
\alias{get_internal_nodes}
\title{Get all internal nodes of a subtree (Helper function)}
\usage{
get_internal_nodes(node, nodes = list())
}
\arguments{
\item{node}{Root of the subtree}

\item{nodes}{List just used for recursion}
}
\value{
List of Node objects
}
\description{
Recursively collects all internal (non-leaf) nodes of a subtree
This helper is used when searching for the weakest link during
cost-complexity pruning
}
\examples{
get_internal_nodes(cart@ref$root)
}
