% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_complexity_pruning.R
\name{get_node_indices}
\alias{get_node_indices}
\title{Find indices of data rows that reach a node given}
\usage{
get_node_indices(root, node, X, indices = 1:nrow(X))
}
\arguments{
\item{root}{Root of the tree}

\item{node}{Target node}

\item{X}{Data matrix}

\item{indices}{Unfiltered row indices}
}
\value{
Integer vector of observation indices.
}
\description{
Goes dwon the tree from the root and returns the indices of the
data rows that would reach the specified node
}
\examples{
get_node_indices(root, node, X)
}
