% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node@assign_children}
\alias{node@assign_children}
\title{Assign children to a node}
\arguments{
\item{node}{A node object, to which children are assigned}

\item{child1}{A node object, to assign to node}

\item{child2}{(Optional) A further node object, to assign to node}
}
\value{
void
}
\description{
Assign children to a node
}
\examples{
assign_children(node,	child1,	child2)
assign_children(node,	child)
}
