% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node@assign_right_child}
\alias{node@assign_right_child}
\title{Assign right child to a node}
\arguments{
\item{node}{A node object}

\item{child2be}{A node object}
}
\value{
void
}
\description{
Assign right child to a node
}
\examples{
assign_right_child(node,	child)
}
