% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@get_left_child}
\alias{Node@get_left_child}
\title{Get the left child of the node}
\arguments{
\item{node}{A node object}

\item{new_node}{A node object}
}
\value{
node@left_node		A node object corresponding to the left child

node@left_node			A node object corresponding to the left child
}
\description{
Get the left child of the node

Set the left child of the node
}
\examples{
get_left_child(node)
set_left_child(node)
}
