% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@get_parent}
\alias{Node@get_parent}
\title{Get parent node of target node}
\arguments{
\item{node}{A node object}
}
\value{
node@parent_node	A node object
}
\description{
Get parent node of target node
}
\examples{
get_parent(node)
}
