% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@get_right_child}
\alias{Node@get_right_child}
\title{Get the right child of the node}
\arguments{
\item{node}{A node object}
}
\value{
node@right_node		A node object corresponding to the right child
}
\description{
Get the right child of the node
}
\examples{
get_right_child(node)
}
