% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@get_value}
\alias{Node@get_value}
\title{Get value of target node. If the node is a leaf, then leaf_value is returned, otherwise the split value is returned.}
\arguments{
\item{node}{A node object}
}
\value{
node@value			A double corresponding to the node's value
}
\description{
Get value of target node. If the node is a leaf, then leaf_value is returned, otherwise the split value is returned.
}
\examples{
get_value(node)
}
