% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node@remove_child}
\alias{node@remove_child}
\title{Remove child from node}
\arguments{
\item{node}{A (child) node object}

\item{return}{void}
}
\description{
Remove child from node
}
\examples{
remove_child(node,	child)
}
