% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@route}
\alias{Node@route}
\title{Decision making method for routing children depending on the node}
\arguments{
\item{node}{Node Object}

\item{obs}{Numeric value vector}
}
\value{
void
}
\description{
Decision making method for routing children depending on the node
}
\examples{
route(node)
}
