% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@set_parent}
\alias{Node@set_parent}
\title{Set the parent node of the target to a new node object}
\arguments{
\item{node}{Target node object}

\item{new_parent_node}{New node object}
}
\description{
Set the parent node of the target to a new node object
}
\examples{
set_parent(target,	new_parent)
set_parent(root,	new_parent)
}
