% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@set_right_child}
\alias{Node@set_right_child}
\title{Set the right child of the node}
\arguments{
\item{node}{A node object}

\item{new_node}{A node object}
}
\value{
node@right_node			A node object corresponding to the right child
}
\description{
Set the right child of the node
}
\examples{
set_right_child(node)
}
