% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@set_value}
\alias{Node@set_value}
\title{Set value of target node}
\arguments{
\item{node}{A node object}

\item{value}{A double value to set}
}
\value{
void				Void
}
\description{
Set value of target node
}
\examples{
set_value(node)
}
