% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{RandomForest@compute_importance}
\alias{RandomForest@compute_importance}
\title{Compute feature importance by split frequency}
\arguments{
\item{rf}{A RandomForest object}
}
\value{
named	Numeric vector of length d, sums to 1
}
\description{
Compute feature importance by split frequency
}
\examples{
compute_importance(rf)
}
