% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{RandomForest@fit_rf}
\alias{RandomForest@fit_rf}
\title{Fit the Random Forest to training data}
\arguments{
\item{rf}{A RandomForest object}

\item{X}{A numeric matrix of features (n x d)}

\item{y}{A numeric vector of target values}
}
\value{
void
}
\description{
Fit the Random Forest to training data
}
\examples{
fit_rf(rf,	X,	y)
}
