% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{RandomForest@get_feature_importance}
\alias{RandomForest@get_feature_importance}
\title{Get feature importance scores}
\arguments{
\item{rf}{A RandomForest object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		Named numeric vector of length d
}\if{html}{\out{</div>}}
}
\description{
Get feature importance scores
}
\examples{
get_feature_importance(rf)
}
