% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{RandomForest@get_tree}
\alias{RandomForest@get_tree}
\title{Get a single tree from the forest}
\arguments{
\item{rf}{A RandomForest object}

\item{b}{Integer index of the tree}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A RandomForestCART object
}\if{html}{\out{</div>}}
}
\description{
Get a single tree from the forest
}
\examples{
get_tree(rf,	1L)
}
