% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{RandomForest@predict_rf}
\alias{RandomForest@predict_rf}
\title{Predict using the Random Forest for a single observation}
\arguments{
\item{rf}{A RandomForest object}

\item{x}{A numeric vector of length d}
}
\value{
Double	(regression) or integer (classification)
}
\description{
Predict using the Random Forest for a single observation
}
\examples{
predict_rf(rf,	x)
}
