% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node@get_value}
\alias{Node@get_value}
\title{Get value of target node}
\arguments{
\item{node}{A node object}
}
\value{
node@value			A double corresponding to the node's value
}
\description{
Get value of target node
}
\examples{
get_value(node)
}
